Move session mutex and shared object classes to dedicated files, to allow for use in mmap client
