+++ uhal/uhal/src/common/ProtocolPCIe.cpp
@@ -285,7 +285,7 @@ bool PCIe::File::haveLock() const
 void PCIe::File::lock()
 {
   if ( flock(mFd, LOCK_EX) == -1 ) {
-    exception::MutexError lExc;
+    detail::MutexError lExc;
     log(lExc, "Failed to lock device file ", Quote(mPath), "; errno=", Integer(errno), ", meaning ", Quote (strerror(errno)));
     throw lExc;
   }
@@ -301,127 +301,6 @@ void PCIe::File::unlock()
   else
     mLocked = false;
 }
-
-
-
-
-PCIe::RobustMutex::RobustMutex() :
-  mCount(0),
-  mSessionActive(false)
-{
-  pthread_mutexattr_t lAttr;
-
-  int s = pthread_mutexattr_init(&lAttr);
-  if (s != 0) {
-    exception::MutexError lExc;
-    log(lExc, "Error code ", Integer(s), " (", strerror(s), ") returned in mutex attr initialisation");
-    throw lExc;
-  }
-
-  s = pthread_mutexattr_setpshared(&lAttr, PTHREAD_PROCESS_SHARED);
-  if (s != 0) {
-    exception::MutexError lExc;
-    log(lExc, "Error code ", Integer(s), " (", strerror(s), ") returned by pthread_mutexattr_setpshared");
-    throw lExc;
-  }
-
-  s = pthread_mutexattr_setrobust(&lAttr, PTHREAD_MUTEX_ROBUST);
-  if (s != 0) {
-    exception::MutexError lExc;
-    log(lExc, "Error code ", Integer(s), " (", strerror(s), ") returned by pthread_mutexattr_setrobust");
-    throw lExc;
-  }
-
-  s = pthread_mutex_init(&mMutex, &lAttr);
-  if (s != 0) {
-    exception::MutexError lExc;
-    log(lExc, "Error code ", Integer(s), " (", strerror(s), ") returned in mutex initialisation");
-    throw lExc;
-  }
-}
-
-
-PCIe::RobustMutex::~RobustMutex()
-{
-}
-
-
-void PCIe::RobustMutex::lock()
-{
-  int s = pthread_mutex_lock(&mMutex);
-  bool lLastOwnerDied = (s == EOWNERDEAD);
-  if (lLastOwnerDied)
-    s = pthread_mutex_consistent(&mMutex);
-
-  if (s != 0) {
-    exception::MutexError lExc;
-    log(lExc, "Error code ", Integer(s), " (", strerror(s), ") returned when ", lLastOwnerDied ? "making mutex state consistent" : "locking mutex");
-    throw lExc;
-  }
-}
-
-
-void PCIe::RobustMutex::unlock()
-{
-  int s = pthread_mutex_unlock(&mMutex);
-  if (s != 0)
-    log(Error(), "Error code ", Integer(s), " (", strerror(s), ") returned when unlocking mutex");
-}
-
-
-uint64_t PCIe::RobustMutex::getCounter() const
-{
-  return mCount;
-}
-
-
-bool PCIe::RobustMutex::isActive() const
-{
-  return mSessionActive;
-}
-
-
-void PCIe::RobustMutex::startSession()
-{
-  mCount++;
-  mSessionActive = true;
-}
-
-void PCIe::RobustMutex::endSession()
-{
-  mSessionActive = false;
-}
-
-
-
-
-template <class T>
-PCIe::SharedObject<T>::SharedObject(const std::string& aName) :
-  mName(aName),
-  mSharedMem(boost::interprocess::open_or_create, aName.c_str(), 1024, 0x0, boost::interprocess::permissions(0666)),
-  mObj(mSharedMem.find_or_construct<T>(boost::interprocess::unique_instance)())
-{
-}
-
-template <class T>
-PCIe::SharedObject<T>::~SharedObject()
-{
-  // boost::interprocess::shared_memory_object::remove(mName.c_str());
-}
-
-template <class T>
-T* PCIe::SharedObject<T>::operator->()
-{
-  return mObj;
-}
-
-template <class T>
-T& PCIe::SharedObject<T>::operator*()
-{
-  return *mObj;
-}
-
-
 
 
 std::string PCIe::getSharedMemName(const std::string& aPath)
@@ -524,7 +403,7 @@ void PCIe::Flush( )
 
   mDeviceFileHostToFPGA.unlock();
 
-  IPCScopedLock_t lLockGuard(*mIPCMutex);
+  detail::ScopedSessionLock lLockGuard(*mIPCMutex);
   mIPCMutex->endSession();
 }
 
@@ -563,12 +442,12 @@ uint32_t PCIe::getMaxReplySize()
 
 void PCIe::connect()
 {
-  IPCScopedLock_t lLockGuard(*mIPCMutex);
+  detail::ScopedSessionLock lLockGuard(*mIPCMutex);
   connect(lLockGuard);
 }
 
 
-void PCIe::connect(IPCScopedLock_t& aGuard)
+void PCIe::connect(detail::ScopedSessionLock& aGuard)
 {
   // Read current value of session counter when reading status info from FPGA
   // (So that can check whether this info is up-to-date later on, when sending next request packet)
@@ -633,7 +512,7 @@ void PCIe::write(const std::shared_ptr<Buffers>& aBuffers)
   if (not mDeviceFileHostToFPGA.haveLock()) {
     mDeviceFileHostToFPGA.lock();
 
-    IPCScopedLock_t lGuard(*mIPCMutex);
+    detail::ScopedSessionLock lGuard(*mIPCMutex);
     mIPCMutex->startSession();
     mIPCSessionCount++;
 
@@ -651,7 +530,7 @@ void PCIe::write(const std::shared_ptr<Buffers>& aBuffers)
   lDataToWrite.push_back( std::make_pair(reinterpret_cast<const uint8_t*>(&lHeaderWord), sizeof lHeaderWord) );
   lDataToWrite.push_back( std::make_pair(aBuffers->getSendBuffer(), aBuffers->sendCounter()) );
 
-  IPCScopedLock_t lGuard(*mIPCMutex);
+  detail::ScopedSessionLock lGuard(*mIPCMutex);
   mDeviceFileHostToFPGA.write(mIndexNextPage * 4 * mPageSize, lDataToWrite);
   log (Debug(), "Wrote " , Integer((aBuffers->sendCounter() / 4) + 1), " 32-bit words at address " , Integer(mIndexNextPage * 4 * mPageSize), " ... ", detail::PacketFmt(lDataToWrite));
 
@@ -699,7 +578,7 @@ void PCIe::read()
       std::vector<uint32_t> lValues;
       while ( true ) {
         // FIXME : Improve by simply adding fileWrite method that takes uint32_t ref as argument (or returns uint32_t)
-        IPCScopedLock_t lGuard(*mIPCMutex);
+        detail::ScopedSessionLock lGuard(*mIPCMutex);
         mDeviceFileFPGAToHost.read(0, (mXdma7seriesWorkaround ? 8 : 4), lValues);
         lHwPublishedPageCount = lValues.at(3);
         log (Debug(), "Read status info from addr 0 (", Integer(lValues.at(0)), ", ", Integer(lValues.at(1)), ", ", Integer(lValues.at(2)), ", ", Integer(lValues.at(3)), "): ", detail::PacketFmt((const uint8_t*)lValues.data(), 4 * lValues.size()));
@@ -737,7 +616,7 @@ void PCIe::read()
   lNrWordsToRead += 1;
  
   std::vector<uint32_t> lPageContents;
-  IPCScopedLock_t lGuard(*mIPCMutex);
+  detail::ScopedSessionLock lGuard(*mIPCMutex);
   mDeviceFileFPGAToHost.read(4 + lPageIndexToRead * mPageSize, lNrWordsToRead , lPageContents);
   lGuard.unlock();
   log (Debug(), "Read " , Integer(lNrWordsToRead), " 32-bit words from address " , Integer(16 + lPageIndexToRead * 4 * mPageSize), " ... ", detail::PacketFmt((const uint8_t*)lPageContents.data(), 4 * lPageContents.size()));
